Extract JSON serialization/deserialization to it's own class
