+++ ElectronicCash/ElectronicCash.csproj
@@ -65,6 +65,7 @@
     <Compile Include="MoneyOrder.cs" />
     <Compile Include="ActorName.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SerializeCustomerModel.cs" />
     <Compile Include="StreetAddress.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
